% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakThrough.R
\name{breakThrough}
\alias{breakThrough}
\title{Says if there is a connected path between the boundary limits}
\usage{
breakThrough(bi, lims, clust)
}
\arguments{
\item{bi}{List with 2 integer vectors wm and wM. ID of line segments touching the study interval boundary. The result from \link{touchBoundary}.}

\item{lims}{Numeric 2-dimensinal vector with the limits of the study interval.}

\item{clust}{a result from \link[igraph]{graph.adjacency}.}
}
\value{
A logical vector of length equals to the number of clusters. Each component answering the question: There is a connected path from \code{lims[1]} to \code{lims[2]}?
}
\description{
Says if there is a connected path between the boundary limits along a given direction. Two posible directions:vertical or horizontal. Not both.
}
\examples{
n <- 6; xl <- -0.3; xr <- 1.2
set.seed(123) #9
endpts <- data.frame(x0 = runif(n, xl, xr),
y0 = runif(n, xl, xr),
x1 = runif(n, xl, xr),
y1 = runif(n, xl, xr))
plotSegments(segment = endpts)
lims <- c(0,1)
abline(h=lims, v= lims, col = 'lightgray')
te <- touchBoundary(w0 = endpts$x0, w1 = endpts$x1, lims = lims)
library(igraph)
eAdjClust <- clusters(graph.adjacency(incidence.matrix(endpts)))
breakThrough(bi=te, lims=0:1, clust = eAdjClust)
}
