% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixIntervals.R
\name{fixIntervals}
\alias{fixIntervals}
\title{Fix interval to be increasingly ordered}
\usage{
fixIntervals(x0, x1)
}
\arguments{
\item{x0}{Numeric vector. Left value of the interval (pair coordinates).}

\item{x1}{Numeric vector. Right value of the interval (pair coordinates).}
}
\value{
A data.frame with columns \code{xl} and \code{xu}, the lower and upper values. Each row is sorted increasingly.
}
\description{
Fix interval to be increasingly ordered
}
\examples{
xl <- c(-1, 0.3, -4,  2)
xu <- c( 1,   1, -2, -3)
fixIntervals(xl, xu)

}
