% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.check.R
\name{inter.check}
\alias{inter.check}
\title{Verify if 2 line segments intersect}
\usage{
inter.check(a, b, m)
}
\arguments{
\item{a}{ID of the matrix.}

\item{b}{don't know yet.}

\item{m}{Matrix with four columns where the first and last two entries of every row are the cartesian coordinates of a single fracture's endpoints.}
}
\value{
the results.
}
\description{
Verify if 2 line segments intersect
}
\details{
This function determines whether two fractures intersect, using the criteria of
opposing orientations in the two-dimensional case. Since it is specifically desinged
to work with the rest of the functions in the model, it takes as parameter the tags
of the fractures which correspond to the row they occupy in the endpoints matrix, as
well as the matrix itself. It doesn't consider cases of colinearity, since given the
measures defined the probability of this eventuality is zero.
}
