% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{orientation}
\alias{orientation}
\title{Signed orientation}
\usage{
orientation(p, q, r)
}
\arguments{
\item{p}{Numeric two-dimensional vector in cartesian coordinates.}

\item{q}{Numeric two-dimensional vector in cartesian coordinates.}

\item{r}{Numeric two-dimensional vector in cartesian coordinates.}
}
\value{
Eihter 1 (positive orientation) or -1 (negative orientation).
}
\description{
Signed orientation
}
\details{
This function determines the orientation of three points in a plane.
Since these points are assumed to come from a random point process, the
colinearity case is not considered, since the probability of this eventuality
is certainly zero.
}
\examples{
orientation(p=c(0,0), q=c(-1,0), r=c(0,1))
orientation(p=c(0,0), q=c( 1,0), r=c(0,1))
}
\author{
Daniella Ayala (\email{daniellaayala@ciencias.unam.mx}) and Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
