% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchBoundary.R
\name{touchBoundary}
\alias{touchBoundary}
\title{Verify line segment intersection with a simple boundary}
\usage{
touchBoundary(w0, w1, lims = 0:1)
}
\arguments{
\item{w0, w1}{Numeric vector. w stands for either x or y coordinate (of both ends) of the line segments.}

\item{lims}{Numeric 2 dimensional vector.}
}
\value{
List with vectors \code{wm}  and \code{wM} with values where the input touches the boundary.
}
\description{
Does the line segments touch the boundary? Verify line segment intersection with a simple boundary.
}
\examples{
n <- 10; xl <- -0.3; xr <- 1.2
set.seed(123) #9
endpts <- data.frame(x0 = runif(n, xl, xr),
y0 = runif(n, xl, xr),
x1 = runif(n, xl, xr),
y1 = runif(n, xl, xr))
plotSegments(segment = endpts)
lims <- c(0,1)
abline(h=lims, v= lims, col = 'lightgray')
# Check for intersection with vertical lines
te <- touchBoundary(w0=endpts$x0, w1 = endpts$x1, lims = lims)
# Intersection with the leftmost vertical boundary (x=constant)
segments(endpts$x0[te$wm], endpts$y0[te$wm],
endpts$x1[te$wm], endpts$y1[te$wm],
col = 'blue', lwd = 3)
# Intersection with the rightmost vertical boundary (x=constant)
segments(endpts$x0[te$wM], endpts$y0[te$wM],
endpts$x1[te$wM], endpts$y1[te$wM],
col = 'green', lwd = 3)

# Intersection with horizontal lines
tey <- touchBoundary(w0=endpts$y0, w1 = endpts$y1, lims = lims)
# Intersection with the bottom-most horizontal boundary (y=constant)
segments(endpts$x0[tey$wm], endpts$y0[tey$wm],
endpts$x1[tey$wm], endpts$y1[tey$wm],
col = 'pink', lwd = 3)
# Intersection with the uppermost horizontal boundary (y=constant)
segments(endpts$x0[tey$wM], endpts$y0[tey$wM],
endpts$x1[tey$wM], endpts$y1[tey$wM],
col = 'red', lwd = 3)
}
